refactor(ui): move attention dismissal store to its own module
